# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Phone booking system API
  version: 1.0.0
servers:
  - url: http://{hostname}:{port}/booking
    variables:
      hostname:
        description: Host where app is running
        default: localhost
      port:
        enum:
          - '8080'
        default: '8080'
paths:
...
  /phone:
    get:
      summary: Get phone data by its manufacturer and model
      operationId: getByManufacturerAndModel
      parameters:
        - name: manufacturer
          in: query
          schema:
            type: string
          description: Manufacturer name
          required: true
        - name: model
          in: query
          schema:
            type: string
          description: Model
          required: true
      responses:
        '200':
          description: Phone representation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneDto"
        '404':
          description: Phone not found by specified parameters
...
components:
...
  schemas:
...
    PhoneDto:
      required:
        - description
        - id
        - isBooked
        - bookedTime
        - bookedBy
        - technology
        - bands
      properties:
        id:
          type: integer
          format: int64
        isBooked:
          type: boolean
        bookedTime:
          type: string
          format: date-time
        bookedBy:
          type: string
        description:
          type: string
        technology:
          type: string
        bands:
          type: array
          items:
            $ref: "#/components/schemas/BandDto"
    BandDto:
      required:
        - bands
        - frequency
      properties:
        band:
          type: string
        frequency:
          type: string